id-provider bigint 2 long

+++ inge_db_connector/src/main/java/de/mpg/mpdl/inge/db/repository/IdentifierProviderServiceImpl.java
@@ -34,7 +34,7 @@ public class IdentifierProviderServiceImpl {
   @Transactional(propagation = Propagation.REQUIRES_NEW)
   public String getNewId(ID_PREFIX prefix) {
 
-    BigInteger res = (BigInteger) entityManager.createNativeQuery("SELECT current_id FROM id_provider FOR UPDATE;").getSingleResult();
+    Long res = (Long) entityManager.createNativeQuery("SELECT current_id FROM id_provider FOR UPDATE;").getSingleResult();
     entityManager.createNativeQuery("UPDATE id_provider SET current_id=current_id+1;").executeUpdate();
 
     return new StringBuilder(prefix.getPrefix()).append("_").append(res.intValue()).toString();
